requirements.txt
pytorch
pandas
wandb